Stratigraphy/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
dotenv==0.9.9
filelock==3.20.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
google-auth==2.47.0
google-genai==1.59.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
onnx==1.20.0
onnxruntime==1.23.2
opencv-python==4.12.0.88
packaging==25.0
pillow==12.1.0
protobuf==6.33.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycocotools==2.0.11
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
torch==2.9.1
torchvision==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
